teamserver/requirements.txt
alive_progress
requests-file==1.5.1
aiohttp
#aliyun-python-sdk-core==2.13.12
aliyun-python-sdk-kms==2.4.1
flask-sqlalchemy==2.5
SQLAlchemy==1.4.40
#requests==2.29.0
requests>=2.31.0
docker
boto3
aniso8601==8.0.0
#click==7.1.2
click==8.0
Flask
Flask-RESTful==0.3.8
#Flask-SQLAlchemy==2.4.3
#itsdangerous==1.1.0
itsdangerous>=2.0
#Jinja2==2.11.2
jinja2>=2.11.3
#MarkupSafe==1.1.1
MarkupSafe>=2.0
pytz==2020.1
six==1.15.0
Werkzeug
flask_httpauth
waitress==2.1.2
flask-mongoengine
docker~=5.0.3
flask_bcrypt
flask-jwt-extended
#termcolor~=1.1.0
JustifyText~=0.2.1.post3
#requests~=2.25.1
boto3~=1.17.55
moto~=2.2.15
colorama~=0.4.3
#botocore~=1.20.55
botocore
psutil~=5.8.0
crtsh~=0.3.1
mongoengine~=0.23.1
tabulate
prettytable
pygments 
unicode
adal
termcolor
msal
websockets
python-socketio
#operator
waitress
flask_jwt_extended
flask_bcrypt
websocket_server
adal
msal
crtsh
